+++ packages/http-client-java/emitter/src/emitter.ts
@@ -3,7 +3,6 @@ import {
   EmitContext,
   getNormalizedAbsolutePath,
   JSONSchemaType,
-  NoTarget,
   resolvePath,
 } from "@typespec/compiler";
 import { spawn } from "child_process";
@@ -12,6 +11,7 @@ import { dump } from "js-yaml";
 import { dirname } from "path";
 import { fileURLToPath } from "url";
 import { CodeModelBuilder } from "./code-model-builder.js";
+import { logError } from "./utils.js";
 
 export interface EmitterOptions {
   namespace?: string;
@@ -40,6 +40,8 @@ export interface EmitterOptions {
   polling?: any;
 
   "group-etag-headers"?: boolean;
+
+  "enable-subclient"?: boolean;
 
   "advanced-versioning"?: boolean;
   "api-version"?: string;
@@ -78,7 +80,7 @@ const EmitterOptionsSchema: JSONSchemaType<EmitterOptions> = {
 
     "enable-sync-stack": { type: "boolean", nullable: true, default: true },
     "stream-style-serialization": { type: "boolean", nullable: true, default: true },
-    "use-object-for-unknown": { type: "boolean", nullable: true, default: true },
+    "use-object-for-unknown": { type: "boolean", nullable: true, default: false },
 
     // customization
     "partial-update": { type: "boolean", nullable: true, default: false },
@@ -89,6 +91,8 @@ const EmitterOptionsSchema: JSONSchemaType<EmitterOptions> = {
     polling: { type: "object", additionalProperties: true, nullable: true },
 
     "group-etag-headers": { type: "boolean", nullable: true },
+
+    "enable-subclient": { type: "boolean", nullable: true, default: false },
 
     "advanced-versioning": { type: "boolean", nullable: true, default: false },
     "api-version": { type: "string", nullable: true },
@@ -132,7 +136,8 @@ export async function $onEmit(context: EmitContext<EmitterOptions>) {
 
     await promises.mkdir(outputPath, { recursive: true }).catch((err) => {
       if (err.code !== "EISDIR" && err.code !== "EEXIST") {
-        throw err;
+        logError(program, `Failed to create output directory: ${outputPath}`);
+        return;
       }
     });
 
@@ -221,17 +226,9 @@ export async function $onEmit(context: EmitContext<EmitterOptions>) {
       // program.trace("http-client-java", output.stdout ? output.stdout : output.stderr);
     } catch (error: any) {
       if (error && "code" in error && error["code"] === "ENOENT") {
-        const msg = "'java' is not on PATH. Please install JDK 11 or above.";
-        program.trace("http-client-java", msg);
-        program.reportDiagnostic({
-          code: "http-client-java",
-          severity: "error",
-          message: msg,
-          target: NoTarget,
-        });
-        throw new Error(msg);
+        logError(program, "'java' is not on PATH. Please install JDK 11 or above.");
       } else {
-        throw error;
+        logError(program, error.message);
       }
     }