+++ packages/http-client-java/emitter/src/operation-utils.ts
@@ -3,14 +3,8 @@ import {
   SdkHttpOperation,
   SdkLroServiceMetadata,
 } from "@azure-tools/typespec-client-generator-core";
-import { ModelProperty, Operation, Program, Type, Union } from "@typespec/compiler";
-import {
-  HttpOperation,
-  getHeaderFieldName,
-  getPathParamName,
-  getQueryParamName,
-  isStatusCode,
-} from "@typespec/http";
+import { getVisibility, ModelProperty, Operation, Program, Type, Union } from "@typespec/compiler";
+import { HttpOperation, isMetadata } from "@typespec/http";
 import { Client as CodeModelClient, ServiceVersion } from "./common/client.js";
 import { CodeModel } from "./common/code-model.js";
 import { modelIs, unionReferredByType } from "./type-utils.js";
@@ -121,15 +115,8 @@ export function operationRefersUnion(
   return null;
 }
 
-export function isPayloadProperty(program: Program, property: ModelProperty | undefined): boolean {
-  if (property === undefined) {
-    return false;
-  }
-  const headerInfo = getHeaderFieldName(program, property);
-  const queryInfo = getQueryParamName(program, property);
-  const pathInfo = getPathParamName(program, property);
-  const statusCodeInfo = isStatusCode(program, property);
-  return !(headerInfo || queryInfo || pathInfo || statusCodeInfo);
+export function isPayloadProperty(program: Program, property: ModelProperty): boolean {
+  return !isMetadata(program, property) && !getVisibility(program, property)?.includes("none");
 }
 
 export function getServiceVersion(client: CodeModelClient | CodeModel): ServiceVersion {