+++ packages/http-client-java/generator/http-client-generator-test/src/test/java/client/structure/DefaultClientTests.java
@@ -10,27 +10,33 @@ import client.structure.service.QuxBarClient;
 import client.structure.service.QuxClient;
 import client.structure.service.ServiceClientClient;
 import client.structure.service.ServiceClientClientBuilder;
+import client.structure.service.models.ClientType;
 import org.junit.jupiter.api.Test;
 
 public class DefaultClientTests {
 
     private final ServiceClientClient client
-        = new ServiceClientClientBuilder().endpoint("http://localhost:3000").client("default").buildClient();
+        = new ServiceClientClientBuilder().endpoint("http://localhost:3000").client(ClientType.DEFAULT).buildClient();
 
-    private final FooClient client2
-        = new ServiceClientClientBuilder().endpoint("http://localhost:3000").client("default").buildFooClient();
+    private final FooClient client2 = new ServiceClientClientBuilder().endpoint("http://localhost:3000")
+        .client(ClientType.DEFAULT)
+        .buildFooClient();
 
-    private final BarClient client3
-        = new ServiceClientClientBuilder().endpoint("http://localhost:3000").client("default").buildBarClient();
+    private final BarClient client3 = new ServiceClientClientBuilder().endpoint("http://localhost:3000")
+        .client(ClientType.DEFAULT)
+        .buildBarClient();
 
-    private final QuxClient client4
-        = new ServiceClientClientBuilder().endpoint("http://localhost:3000").client("default").buildQuxClient();
+    private final QuxClient client4 = new ServiceClientClientBuilder().endpoint("http://localhost:3000")
+        .client(ClientType.DEFAULT)
+        .buildQuxClient();
 
-    private final QuxBarClient client5
-        = new ServiceClientClientBuilder().endpoint("http://localhost:3000").client("default").buildQuxBarClient();
+    private final QuxBarClient client5 = new ServiceClientClientBuilder().endpoint("http://localhost:3000")
+        .client(ClientType.DEFAULT)
+        .buildQuxBarClient();
 
-    private final BazFooClient client6
-        = new ServiceClientClientBuilder().endpoint("http://localhost:3000").client("default").buildBazFooClient();
+    private final BazFooClient client6 = new ServiceClientClientBuilder().endpoint("http://localhost:3000")
+        .client(ClientType.DEFAULT)
+        .buildBazFooClient();
 
     @Test
     public void testClient() {